% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_function.R
\name{boosting_core}
\alias{boosting_core}
\title{Boosting core function}
\usage{
boosting_core(formula, data, rate, control = 500, control_method = NULL,
  control_parameter = NULL, censoring_type = "right")
}
\arguments{
\item{formula}{a formula object with a response value using the Surv function.}

\item{data}{a data.frame containing all variables specified in the formula.}

\item{rate}{the desired update rate used in the boosting algorithm.}

\item{control}{an integer used as the number of iterations of the boosting algorithm. Default value is 500.}

\item{control_method}{specifies stopping method, options include: cv, num_selected, likelihood, BIC, AIC. Default is NULL, which will use a fixed number of iterations as specified by control.}

\item{control_parameter}{provides the parameter needed for each corresponding control_method option.}
}
\value{
a list containing the vector of coefficients (beta), variable selection matrix that contains the coefficients at each iteration (selection_df), the number of boosting iterations (mstop), and other stopping criteria if applicable to selected method.
}
\description{
This function allows you to use gradient boosting for variable selection.
}
\keyword{boosting}
\keyword{gradient}
