% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_function.R
\name{inference.boosting}
\alias{inference.boosting}
\title{Boosting inference function}
\usage{
inference.boosting(fmla, data)
}
\arguments{
\item{data}{data.frame containing variables included in the formula.}

\item{formula}{a formula object obtained from the summary of the boosting_core function.}
}
\description{
This function provides post selection inference.
}
\keyword{boosting}
\keyword{gradient}
