% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_function.R
\name{modelfit.boosting}
\alias{modelfit.boosting}
\title{Boosting model fit function}
\usage{
modelfit.boosting(x, all_beta = NULL)
}
\arguments{
\item{output}{output from boosting_core function.}
}
\description{
This function provides details of the model fit from the boosting_core function.
}
\keyword{boosting}
\keyword{gradient}
