% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_function.R
\name{plot.boosting}
\alias{plot.boosting}
\title{Boosting plot function}
\usage{
\method{plot}{boosting}(x, type = "frequency")
}
\arguments{
\item{output}{output from the boosting_core function.}
}
\description{
This function allows you to visualize the coefficient paths of the boosting algorithm.
}
\keyword{boosting}
\keyword{gradient}
