% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_function.R
\name{predict.boosting}
\alias{predict.boosting}
\title{Boosting predict function}
\usage{
\method{predict}{boosting}(x, new_data = NULL)
}
\arguments{
\item{new_data}{data.frame used for prediction. Default is NULL and will use data specified for boosting algorithm.}

\item{output}{output from boosting_core function.}
}
\value{
vector of the hazard ratio for each observation relative to the sample average.
}
\description{
This function predicts
}
\keyword{boosting}
\keyword{gradient}
