% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survival_cox.R
\name{simulate_survival_cox}
\alias{simulate_survival_cox}
\title{Simulating survival data}
\usage{
simulate_survival_cox(true_beta, base_hazard = "auto",
  base_hazard_scale = NULL, base_hazard_shape = NULL, num_strata = 10,
  input_strata_size = 50, z_matrix = NULL, cov_structure = "diag",
  block_size = 1, rho = NULL, censor_dist = "unif", censor_const = 5,
  tau = Inf, normalized = F)
}
\arguments{
\item{true_beta}{Contains true parameter values to simulate from.}
}
\value{
a matrix with survival time (time), event indicator (delta), stratification variable (strata_idx), a vector for each variable specified by the true_beta.
}
\description{
This function allows you to simulate stratified survival data.
}
\keyword{boosting}
\keyword{gradient}
