% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_function.R
\name{summary.boosting}
\alias{summary.boosting}
\title{Summary of boosting model selection}
\usage{
\method{summary}{boosting}(x)
}
\arguments{
\item{output}{output from boosting_core function.}
}
\value{
list containing the coefficient vector, number of boosting iterations, and resulting formula from the variable selection.
}
\description{
This function displays the variables selected from a model generated with the boosting_core function.
}
\keyword{boosting}
\keyword{gradient}
