% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting_function.R
\name{inference.boosting}
\alias{inference.boosting}
\title{Boosting inference function}
\usage{
inference.boosting(x)
}
\arguments{
\item{x}{output from boosting_core function.}
}
\description{
This function provides post selection inference.
}
\examples{
data <- simulate_survival_cox(true_beta=c(1,1,1,1,1,0,0,0,0,0))
formula <- as.formula("Surv(time,delta) ~ strata(strata_idx) + V1 + V2 + 
V3 + V4 + V5 + V6 + V7 + V8 + V9 + V10" )
boosting.output <- boosting_core(formula, data, rate=0.1, num_iter=500)
inference.boosting(boosting.output)

}
\keyword{boosting}
\keyword{gradient}
